remove cernerFacilitiesAlert from Medical records domain pages
