+++ src/applications/mhv-medical-records/tests/e2e/accelerated/allergies/view-allergies.accelerated.cypress.spec.js
@@ -23,8 +23,10 @@ describe('Medical Records Accelerated Allergies', () => {
 
     cy.injectAxeThenAxeCheck();
 
-    // OH user, in the "accelerated" feature toggle group should see blue alert
-    cy.get('[data-testid="cerner-facilities-info-alert"]').should('be.visible');
+    // Alert removed from Allergies page — assert it does not render
+    cy.get('body')
+      .find('[data-testid="cerner-facilities-info-alert"]')
+      .should('not.exist');
 
     // Verify unified allergy format is displayed
     cy.get('[data-testid="record-list-item"]').should(
