+++ src/applications/mhv-medical-records/tests/e2e/accelerated/conditions/view-conditions-list.cypress.spec.js
@@ -23,8 +23,10 @@ describe('Medical Records View Conditions', () => {
 
     cy.injectAxeThenAxeCheck();
 
-    // OH user, in the "accelerated" feature toggle group should see blue alert
-    cy.get('[data-testid="cerner-facilities-info-alert"]').should('be.visible');
+    // Alert removed from Conditions page — assert it does not render
+    cy.get('body')
+      .find('[data-testid="cerner-facilities-info-alert"]')
+      .should('not.exist');
 
     const CARDS_PER_PAGE = 10;
     cy.get(':nth-child(4) > [data-testid="record-list-item"]').should(
